# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SEN Migration Dashboard API
  description: |
    REST API for the SEN Migration Platform.
    Provides migration status, object progress, rule analysis,
    reconciliation, test scenario management, forensic extraction,
    process mining, data export, and migration planning.
  version: 1.0.0
  contact:
    name: SEN Team
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4004
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: sapconnect.example.com
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/forensic/modules/{id}:
    get:
      tags: [Forensic]
      summary: Get module detail
      operationId: getForensicModule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module report
          content:
            application/json:
              schema:
                type: object
        '404':
          description: No data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/forensic/processes/{id}:
    get:
      tags: [Forensic]
      summary: Get process detail
      operationId: getForensicProcess
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Process detail
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/forensic/config/{module}:
    get:
      tags: [Forensic]
      summary: Get module configuration interpretations
      operationId: getForensicConfig
      parameters:
        - name: module
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration interpretations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /api/forensic/extract:
    post:
      tags: [Forensic]
      summary: Trigger forensic extraction
      operationId: triggerExtraction
      responses:
        '200':
          description: Extraction started
          content:
            application/json:
              schema:
                type: object
        '409':
          description: Extraction already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/signavio/models/{id}:
    get:
      tags: [Signavio]
      summary: Get Signavio model by ID
      operationId: getSignavioModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Process model detail
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/signavio/models/{id}/bpmn:
    get:
      tags: [Signavio]
      summary: Export model as BPMN XML
      operationId: exportSignavioBpmn
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: BPMN XML export
          content:
            application/json:
              schema:
                type: object
                properties:
                  modelId:
                    type: string
                  revisionId:
                    type: string
                  bpmn:
                    type: string
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/signavio/models/{id}/parse:
    post:
      tags: [Signavio]
      summary: Parse BPMN into configuration steps
      operationId: parseSignavioBpmn
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bpmnXml:
                  type: string
      responses:
        '200':
          description: Parsed configuration steps
          content:
            application/json:
              schema:
                type: object
                properties:
                  modelId:
                    type: string
                  configSteps:
                    type: array
                    items:
                      type: object
                  complexity:
                    type: object
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/signavio/search:
    get:
      tags: [Signavio]
      summary: Search Signavio models
      operationId: searchSignavioModels
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Required when API_KEY environment variable is set.
        Pass via X-API-Key header or apiKey query parameter.
        Exempt paths: /health, /ready, /metrics
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...